refactor ts verification into separate files
