+++ sdk/confidentialledger/azure-codetransparency/tests/test_recorded_codetransparency_client.py
@@ -1,9 +1,6 @@
 # pylint: disable=line-too-long,useless-suppression
 # cSpell:ignore operationid
-import hashlib
-import os
-import time
-from typing import Dict, List, Union, Any
+from typing import Dict, List, Union
 from urllib.parse import urlparse
 from azure.core import exceptions
 import pytest
